backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
langchain==0.1.0
langchain-openai==0.0.2
playwright==1.40.0
Pillow==10.1.0
